+++ src/frontend/environments/environment.beta.ts
@@ -1,4 +1,5 @@
 import {AppEnvironment} from './app-environment';
+import {DataMode} from '../shared/commons/model/datamode.enum';
 
 export const environment: AppEnvironment = {
     production: true,
@@ -10,5 +11,7 @@ export const environment: AppEnvironment = {
     allowAutoPlay: false,
     backendApiBaseUrl: 'http://localhost:4101/api/v1/',
     cookieLawSeenName: 'cookieLawSeenV20180525',
-    trackingProviders: [] // Angulartics2Piwik
+    trackingProviders: [], // Angulartics2Piwik
+    startDataMode: DataMode.BACKEND,
+    availableDataModes: [DataMode.BACKEND]
 };